+++ app/dashboards/building_dashboard.rb
@@ -13,7 +13,8 @@ class BuildingDashboard < Administrate::BaseDashboard
   ATTRIBUTE_TYPES = {
     suites: Field::HasMany,
     id: Field::Number,
-    name: Field::String,
+    full_name: Field::String,
+    abbreviation: Field::String,
     created_at: Field::DateTime,
     updated_at: Field::DateTime
   }.freeze
@@ -24,7 +25,8 @@ class BuildingDashboard < Administrate::BaseDashboard
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
   COLLECTION_ATTRIBUTES = %i(
-    name
+    full_name
+    abbreviation
     suites
   ).freeze
 
@@ -33,7 +35,8 @@ class BuildingDashboard < Administrate::BaseDashboard
   SHOW_PAGE_ATTRIBUTES = %i(
     suites
     id
-    name
+    full_name
+    abbreviation
     created_at
     updated_at
   ).freeze
@@ -43,7 +46,8 @@ class BuildingDashboard < Administrate::BaseDashboard
   # on the model's form (`new` and `edit`) pages.
   FORM_ATTRIBUTES = %i(
     suites
-    name
+    full_name
+    abbreviation
   ).freeze
 
   # Overwrite this method to customize how buildings are displayed