+++ db/migrate/20191108171248_add_abbreviation_to_building.rb
@@ -1,0 +1,6 @@
+class AddAbbreviationToBuilding < ActiveRecord::Migration[5.1]
+  def change
+    rename_column(:buildings, :name, :full_name)
+    add_column(:buildings, :abbreviation, :string, null: true)
+  end
+end