+++ spec/features/buildings/building_creation_spec.rb
@@ -5,11 +5,11 @@ require 'rails_helper'
 RSpec.feature 'Building Creation' do
   before { log_in create(:admin) }
   it 'succeeds' do
-    name = 'Silliman'
+    full_name = 'Silliman'
     navigate_to_view
-    fill_in 'building_name', with: name
+    fill_in 'Full name', with: full_name
     click_on 'Create'
-    expect(page).to have_content(name)
+    expect(page).to have_content(full_name)
   end
   it 'redirects to /new on failure' do
     visit 'buildings/new'
@@ -17,9 +17,23 @@ RSpec.feature 'Building Creation' do
     expect(page).to have_content('errors')
   end
 
+  it 'succeeds with an abbreviation' do
+    full_name = 'Silliman'
+    abbreviation = 'SI'
+    navigate_to_view
+    create_building_with_abbreviation(full_name, abbreviation)
+    expect(page).to have_content(full_name)
+  end
+
   def navigate_to_view
     visit root_path
     click_on 'Inventory'
     click_on 'Add new building'
+  end
+
+  def create_building_with_abbreviation(full_name, abbreviation)
+    fill_in 'Full name', with: full_name
+    fill_in 'Abbreviation', with: abbreviation
+    click_on 'Create'
   end
 end