+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 * Add ability to reassign suite in the suite selection or results phase ([#969] (https://gitlab.com/yale-sdmp/vesta/issues/969)).
 * Add masquerading to allow admins to impersonate other users ([#147](https://gitlab.com/yale-sdmp/vesta/issues/147)).
 * Add information about pending group requests to user dashboard ([#696](https://gitlab.com/yale-sdmp/vesta/issues/696)).
+* Add building abbreviations ([#570](https://gitlab.com/yale-sdmp/vesta/issues/570)).
 
 ### Changed
 * Sort groups in results and suite selection tables by lottery number ([#968](https://gitlab.com/yale-sdmp/vesta/issues/968)).